Set1: Complete challenge6
Complete challenge 6 which required calculating the hamming distance
between different slices of the ciphertext. Deriving the key length by
getting the keylength that yields the minimum hamming distance for the
first two slices of the ciphertext did not yield the correct key for me,
so I wrote a method that will computer the average hamming length for
any combination of the first n slices of ct that is the specified
length.