Fix PKCS#7 padding and CBC encrypt
If a plaintext was aligned on a block barrier NO padding would be added.
Because our decrypt function didn't remove it it wasn't an issue. Once I
added code to strip the padding it failed on these texts, so I had to
fix the padding function. This then broke CBC encrypt because each block
was padded with an extra 16*"\x16" block.